requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
bcrypt==4.3.0
PyJWT==2.10.1
pyotp==2.9.0
qrcode[pil]==8.0
python-dotenv==1.1.0
